refactor: repackage as aggregates
