+++ README.md
@@ -9,6 +9,7 @@
 - LotteryBuyingMoney
     - [x] 로또 한 장의 가격은 1000원이다.
     - [x] [제약조건] 로또 구매 가격은 로또 가격에 자연수를 곱한 값이어야 한다.
+    - [ ] 돈 객체를 composition 으로 가지고 있고, Money 라는 인터페이스를 구현하도록 만드는 것은 어떨까?
 - LotteryNumber
     - [x] 로또 번호는 개수가 제한되어있고 값이 변하지 않는 불변객체이므로 캐싱한다.
     - [x] [제약조건] 로또 번호는 1 이상 45 이하이다.
