+++ src/main/java/lotto/domain/lottery/InvalidLotteryNumberException.java
@@ -1,4 +1,4 @@
-package lotto.domain;
+package lotto.domain.lottery;
 
 public class InvalidLotteryNumberException extends RuntimeException {
     private static final String ERROR_MSG = "로또 번호는 1 이상 45 이하여야 합니다.";
