+++ src/main/java/lotto/domain/lottery/Lottery.java
@@ -1,4 +1,4 @@
-package lotto.domain;
+package lotto.domain.lottery;
 
 import java.util.List;
 import java.util.stream.Collectors;
