+++ src/main/java/lotto/domain/lottery/LotteryNumber.java
@@ -1,4 +1,4 @@
-package lotto.domain;
+package lotto.domain.lottery;
 
 import java.util.HashMap;
 import java.util.Map;
