+++ src/main/java/lotto/domain/lottery/LotteryNumbers.java
@@ -1,4 +1,4 @@
-package lotto.domain;
+package lotto.domain.lottery;
 
 import java.util.ArrayList;
 import java.util.List;
