+++ src/main/java/lotto/domain/money/Money.java
@@ -1,4 +1,4 @@
-package lotto.domain;
+package lotto.domain.money;
 
 import java.math.BigInteger;
