+++ src/main/java/lotto/domain/money/MoneyLowerThanZeroException.java
@@ -1,4 +1,4 @@
-package lotto.domain;
+package lotto.domain.money;
 
 public class MoneyLowerThanZeroException extends RuntimeException {
     private static final String ERROR_MSG = "돈은 0 이상의 정수여야 합니다.";
