+++ src/test/java/lotto/domain/LotteryBuyingMoneyTest.java
@@ -1,5 +1,7 @@
 package lotto.domain;
 
+import lotto.domain.lottery.InvalidLotteryBuyingMoneyException;
+import lotto.domain.money.LotteryBuyingMoney;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
