+++ src/test/java/lotto/domain/LotteryNumberTest.java
@@ -1,5 +1,7 @@
 package lotto.domain;
 
+import lotto.domain.lottery.InvalidLotteryNumberException;
+import lotto.domain.lottery.LotteryNumber;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
