+++ src/test/java/lotto/domain/LotteryNumbersTest.java
@@ -1,5 +1,9 @@
 package lotto.domain;
 
+import lotto.domain.lottery.DuplicatedLotteryNumbersException;
+import lotto.domain.lottery.InvalidLotteryNumbersException;
+import lotto.domain.lottery.LotteryNumber;
+import lotto.domain.lottery.LotteryNumbers;
 import org.assertj.core.util.Lists;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
